backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
python-multipart==0.0.6
